Moved API code to src folder and converted to 4 spaces
