--- api/models/Doctor.js
@@ -1,15 +1,0 @@
-const mongoose = require('mongoose')
-
-const doctorSchema = new mongoose.Schema({
-  firstName: String,
-  lastName: String,
-  email: String,
-  clinic: String,
-  userId: Number,
-  password: String,
-  patientIds: [{
-    type: Number
-  }]
-})
-
-module.exports = mongoose.model('Doctor', doctorSchema)
