+++ api/package.json
@@ -2,9 +2,9 @@
   "name": "docpocket-api",
   "version": "1.0.0",
   "description": "api for mobile docpocket app",
-  "main": "server.js",
+  "main": "src/server.js",
   "scripts": {
-    "start": "./node_modules/nodemon/bin/nodemon.js ./server.js",
+    "start": "nodemon .",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
